controllers created

+++ src/main/java/com/etiya/ecommerceDemo/repositories/abstracts/ProductReviewDao.java
@@ -3,5 +3,5 @@ package com.etiya.ecommerceDemo.repositories.abstracts;
 import com.etiya.ecommerceDemo.entities.concretes.ProductReview;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductReviewDao extends JpaRepository<ProductReview,Integer> {
+public interface ProductReviewDao extends JpaRepository<ProductReview,Long> {
 }
