+++ src/main/java/com/etiya/ecommerceDemo/repositories/abstracts/UserDao.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface UserDao extends JpaRepository<User,Integer> {
+public interface UserDao extends JpaRepository<User,Long> {
     @Query(value = "SELECT * FROM users", nativeQuery = true)
     List<User> findAllUser();
 }
